% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisAbility.R
\name{axis.Ability}
\alias{axis.Ability}
\title{Ability axis acquisiton}
\usage{
axis.Ability(n, time, money, routine, peff, meff, socialdev, est.time,
  est.money, est.routine, est.peff, est.meff, est.socialdev, full.estimation)
}
\arguments{
\item{n}{The number of observation on the dataset or to estimate}

\item{time}{vector of values for time}

\item{money}{vector of values for money}

\item{routine}{vector of values for routine}

\item{peff}{vector of values for physical effort}

\item{meff}{vector of values for mental effort}

\item{socialdev}{vector of values for social deviance}

\item{est.time}{estimates a vector of values for time based on the n size}

\item{est.money}{estimates  a vector of values for money based on the n size}

\item{est.routine}{estimates  a vector of values for routine based on the n size}

\item{est.peff}{estimates a vector of values for physical effort based on the n size}

\item{est.meff}{estimates a vector of values for mental effort based on the n size}

\item{est.socialdev}{estimates a vector of values for social deviance based on the n size}

\item{full.estimation}{performs a estimation for all variables based on the n provided.
If any variable is informed it will be ignored.}
}
\value{
This function returns the Ability value for each observation on
the dataset informed or estimated
}
\description{
This function provides the value for the Ability axis
}
\details{
The domain of values is between 0 to 1. In order to obtain a valid
result, when a vector of values is informed, not estimated, it must be on
the range 0,1

When a vector of values is informed and another variable is estimated, the n
must be equal to all variables in order to calculate.

The amount of positive variables must be equal to negative variables.
}
\examples{
a.estimated <- axis.Ability(n=450, full.estimation=TRUE)
a <- axis.Ability(n=3000, time= df[1:1000,1], est.socialdev=TRUE)

}
\references{
Marcus Guimaraes <guimaraesmvf@gmail.com>
}

