% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremeMA.R
\name{extremMA.high}
\alias{extremMA.high}
\title{Extreme High Motivation and Ability Value}
\usage{
extremMA.high(a, m, main)
}
\arguments{
\item{a}{vector of Ability values}

\item{m}{vector of Motivation values}

\item{main}{plot title}
}
\description{
This function identifies all extreme high Motivation and Ability values
on the dataset
}
\examples{
extremeMA.high(a.val, m.val)
extremeMA.low(a.val, m.val)

}
\references{
Marcus Guimaraes <guimaraesmvf@gmail.com>
}

