% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTrigger.R
\name{findTriggers}
\alias{findTriggers}
\title{Trigger Location and dataset values identification}
\usage{
findTriggers(a, m, dyn, fbm, zone, type, main)
}
\arguments{
\item{a}{vector of Ability values}

\item{m}{vector of Motivation values}

\item{dyn}{plots the dynamic threshold line}

\item{fbm}{plots the theoretical fbm threshold line}

\item{zone}{identifies all triggers zones}

\item{type}{highlights the dataset values for each trigger type:
"spark", "facilitator", "signal", "all"}

\item{main}{plot title}
}
\description{
This function identifies all values on the dataset and highlights
to the correspondent triggers
}
\examples{
findTriggers(a.val, m.val, zone=T)
findTriggers(a.val, axis.Motivation(n=100, full.estimate=T), type="spark", dyn=T)
findTriggers(a.val, m.val, zone=T, type="all", dyn=T)

}
\references{
Marcus Guimaraes <guimaraesmvf@gmail.com>
}

