% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMA.R
\name{plotMA}
\alias{plotMA}
\title{Motivation and Ability Plot}
\usage{
plotMA(a, m, id, samplePlot, harmean, dynamic.threshold, fbm.threshold, main)
}
\arguments{
\item{a}{vector of Ability values}

\item{m}{vector of Motivaton Values}

\item{id}{identifies the Motivation and Ability for a given database register and
informs each trigger type is best recommended}

\item{samplePlot}{example plot using basic options and auto generation via
axis.Ability and axis.Motivation functions}

\item{harmean}{plots the harmonic mean of the dataset}

\item{dynamic.threshold}{plots a dynamic threshold line allocated on the dataset
central tendency}

\item{fbm.threshold}{plots the original (fixed) threshold line proposed on the fbm}

\item{main}{plot title}
}
\description{
This function generates the Motivation and Ability plot for a dataset
according to the FBM. Some options may be used in order to obtain maximum visual information
}
\examples{
plotMA(abil, motiv)
plotMA(samplePlot=TRUE)
plotMA(axis.Ability(n=100,full.estimation=T), axis.Motivation(n=100, full.estimation=T))
plotMA(axis.Ability(n=30, time=abil[1:30,1], est.socialdev=T), motiv[1:30,1:2])
plotMA(aVal, mVal, id=100, dynamic.threshold=T)
plotMA(axis.Ability(n=150, full.estimation=T), axis.Motivation(n=150, full.estimation=T), id=50)

}
\references{
Marcus Guimaraes <guimaraesmvf@gmail.com>
}

