% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeMA.R
\name{rangeMA}
\alias{rangeMA}
\title{rangeMA **<beta>**}
\usage{
rangeMA(a, m)
}
\arguments{
\item{a}{vector of Ability values}

\item{m}{vector of Motivation values}

\item{plot}{if TRUE a visual information about the data classification will be
displayed.}
}
\description{
This function classifies all observation related to its degree
of pertinence based on the dataset harmonic deviation. Observation will be
classified as:

High (observations above 2 harmonic deviations)

Low (observations below 2 harmonic deviations)

Moderate (observations on the range between 1 and 2 harmonic deviations)

Regular (observations on the range between one harmonic.deviation)
}
\examples{
rangeMA(aValues, mValues, plot=TRUE)
rangeMA(abilityVector, axis.Motivation(n=100, full.estimate=T), plot=T)
rangeMA(abil, motiv)
}
\references{
Marcus Guimaraes <guimaraesmvf@gmail.com>
}

