% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggerStatus.R
\name{trigger.status}
\alias{trigger.status}
\title{Dataset proportion on each trigger zone}
\usage{
trigger.status(a, m, plot)
}
\arguments{
\item{a}{vector of Ability values}

\item{m}{vector of Motivation values}

\item{plot}{if TRUE, plots a visual information for all triggers.
Outputs the function findTrigger(a, m, dyn=T, type="all")}
}
\description{
This function informs the percentage of users on the given dataset
that belongs to each specific trigger type. Also, provides a graphical information.
}
\examples{
trigger.status(a.est, m.est)
trigger.status(axis.Ability(n=50, full.estimation=T), m.est, plot=T)

}
\references{
Marcus Guimaraes <guimaraesmvf@gmail.com>
}

